Replace sequences with identity
